Set up simple WASM tests
