+++ ffi/capi/wasm-test/package.json
@@ -1,0 +1,20 @@
+{
+  "ava": {
+    "files": [ "*.mjs" ],
+    "require": [
+      "esm"
+    ],
+    "nodeArguments": [
+			"--experimental-wasm-bigint"
+		]
+  },
+  "scripts": {
+    "test": "ava"
+  },
+  "devDependencies": {
+    "ava": "^3.15.0"
+  },
+  "dependencies": {
+    "esm": "^3.2.25"
+  }
+}
